Terrain: Create Unit Test
