+++ src/Utilities/QGCLoggingCategory.cc
@@ -77,6 +77,7 @@ void QGCLoggingCategoryRegister::setFilterRulesFromSettings(const QString& comma
     }
 
     filterRules += "*Log.debug=false\n";
+    filterRules += "qgc.*.debug=false\n";
 
     // Set up filters defined in settings
     foreach (QString category, _registeredCategories) {
